content-security-policy/navigation-to/navigation-to-allowed-link-click.html

<!DOCTYPE html>

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  var t = async_test("Test that the child iframe navigation is blocked");
  window.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.result, 'success');
  });
</script>
<iframe src="support/link_click_navigation.sub.html?csp=navigation-to%20%27self%27&target=post_message_to_top.html">
</body>
